+++ packages/utils/src/toArray.ts
@@ -1,0 +1,7 @@
+export default function toArray<T>(value: T | T[]): T[] {
+  if (!value) {
+    return [];
+  }
+
+  return Array.isArray(value) ? value : [value];
+}