+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-- [aesthetic](./packages/core/CHANGELOG.md)
+- [aesthetic](./packages/core-legacy/CHANGELOG.md)
 - [aesthetic-adapter-aphrodite](./packages/adapter-aphrodite/CHANGELOG.md)
 - [aesthetic-adapter-css-modules](./packages/adapter-css-modules/CHANGELOG.md)
 - [aesthetic-adapter-fela](./packages/adapter-fela/CHANGELOG.md)
 - [aesthetic-adapter-jss](./packages/adapter-jss/CHANGELOG.md)
 - [aesthetic-adapter-typestyle](./packages/adapter-typestyle/CHANGELOG.md)
-- [aesthetic-react](./packages/react/CHANGELOG.md)
-- [aesthetic-utils](./packages/utils/CHANGELOG.md)
+- [aesthetic-react](./packages/react-legacy/CHANGELOG.md)
+- [aesthetic-utils](./packages/utils-legacy/CHANGELOG.md)